pyproject.toml
[project]
name = "rag-eval-harness"
version = "0.1.0"
description = "Benchmark RAG retrieval strategies head-to-head with RAGAS metrics. Model-agnostic: swap LLMs and embeddings via config."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Sai Vikas Reddy Yeddulamala", email = "saivikas.y@zohomail.com" }
]
keywords = ["rag", "llm", "evaluation", "ragas", "benchmark", "retrieval"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies — works out of the box with Groq + local embeddings
dependencies = [
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-community>=0.3.0",
  "langchain-groq>=0.2.0",
  "langchain-huggingface>=0.1.0",
  "ragas>=0.2.0",
  "faiss-cpu>=1.7.4",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0.0",
  "groq>=0.11.0",
  "cohere>=5.0.0",
  "datasets>=2.20.0",
  "pandas>=2.0.0",
  "plotly>=5.22.0",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.0",
  "rich>=13.0.0",
  "click>=8.1.0",
  "pydantic>=2.0.0",
  "tiktoken>=0.7.0",
  "tqdm>=4.66.0",
]

# Optional provider extras — install only what you need
[project.optional-dependencies]
openai = [
  "openai>=1.40.0",
  "langchain-openai>=0.2.0",
]
anthropic = [
  "anthropic>=0.34.0",
  "langchain-anthropic>=0.2.0",
]
ollama = [
  "ollama>=0.3.0",
  "langchain-ollama>=0.2.0",
]
google = [
  "google-generativeai>=0.8.0",
  "langchain-google-genai>=2.0.0",
]
# Install everything
all = [
  "rag-eval-harness[openai,anthropic,ollama,google]",
]

[project.scripts]
rag-eval = "rag_eval.cli:main"

[project.urls]
Homepage = "https://github.com/saivikasreddy717/rag-eval-harness"
Repository = "https://github.com/saivikasreddy717/rag-eval-harness"
Issues = "https://github.com/saivikasreddy717/rag-eval-harness/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_eval"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    # RAGAS 0.4.x deprecates the LangChain wrapper in favour of llm_factory.
    # We keep using the wrapper because it is the only provider-agnostic path
    # that supports Groq / Anthropic / Ollama out of the box.  Will migrate
    # when RAGAS ships a proper multi-provider factory.
    "ignore:LangchainLLMWrapper is deprecated:DeprecationWarning",
    "ignore:LangchainEmbeddingsWrapper is deprecated:DeprecationWarning",
    # FAISS ships SWIG C-extension types that lack __module__. Harmless.
    "ignore:builtin type Swig:DeprecationWarning",
    "ignore:builtin type swigvarlink:DeprecationWarning",
]
